Big code synchronization and refactoring.

+++ trunk/RTSPProxy-Core/src/main/java/rtspproxy/lib/number/UnsignedShort.java
@@ -74,12 +74,12 @@ public final class UnsignedShort extends UnsignedNumber {
 
 	@Override
 	public double doubleValue() {
-		return (double) value;
+		return value;
 	}
 
 	@Override
 	public float floatValue() {
-		return (float) value;
+		return value;
 	}
 
 	@Override
@@ -124,7 +124,7 @@ public final class UnsignedShort extends UnsignedNumber {
 	
 	@Override
 	public int hashCode() {
-		return (int)value;
+		return value;
 	}
 
 	@Override
